requirements.txt
espn-api>=0.30.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
pydantic>=1.8.0
pytest>=6.2.0
rich>=10.0.0